api/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.12
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
google-generativeai==0.8.3
langchain-pinecone==0.2.0
pinecone-client==5.0.1
pypdf==5.0.1
pillow==10.4.0
python-dotenv==1.0.1
tiktoken==0.7.0